Added stop tokens, modified generate practice to work with new model
